requirements.txt
# KICE Math Problem Management System Dependencies

# Supabase
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Web Server (FastAPI)
fastapi>=0.104.0
uvicorn>=0.24.0

# Google Drive API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# PDF processing
PyMuPDF>=1.23.0
Pillow>=10.0.0

# OCR (optional - for hybrid split verification)
pytesseract>=0.3.10

# Notion API
notion-client>=2.0.0

# Scheduler (optional)
schedule>=1.2.0

# Utilities
python-dateutil>=2.8.2
